Remove server state and load data at page load
Using server state is preverable, but the server state that was
used is only generated when the site is being build/developed.
The data that was used for server state was not static, but editable
by the user. Which means that, after a refresh, the old data would
reappear. The server state is now loaded at page start. The preferable
other option would be to load the data dynamically, on the server
which puts it as a string in the index.html.